refactor(android): extract shared API exception mapping
CheckInRepository and EditionRepository each carried an identical
try/catch block mapping HttpException/SSLPeerUnverifiedException/
IOException to ApiException, with ApiException itself only reachable
from EditionRepository via same-package visibility. Move ApiException,
UnauthorizedException, and the mapping logic into a shared runApiCall
so there's one place to update as more repositories are added.

Also reword the certificate-pin failure message to mention
"certificate" explicitly, since the unit test asserts on that word and
it's clearer for anyone reading logs than "secure connection".
