+++ android/app/src/main/kotlin/be/champagnefestival/android/data/repository/CheckInRepository.kt
@@ -6,8 +6,6 @@ import be.champagnefestival.android.data.model.CheckInLookupRequest
 import be.champagnefestival.android.data.model.CheckInOut
 import be.champagnefestival.android.data.model.CheckInRequest
 import retrofit2.HttpException
-import java.io.IOException
-import javax.net.ssl.SSLPeerUnverifiedException
 
 interface CheckInRepository {
     suspend fun lookupRegistration(
@@ -33,19 +31,25 @@ class DefaultCheckInRepository(
     override suspend fun lookupRegistration(
         id: String,
         token: String,
-    ): Result<CheckInGuestOut> = runApiCall { apiService.lookupCheckIn(id = id, body = CheckInLookupRequest(token)) }
+    ): Result<CheckInGuestOut> =
+        runApiCall(mapHttpException = ::mapAuthAwareException) {
+            apiService.lookupCheckIn(id = id, body = CheckInLookupRequest(token))
+        }
 
     override suspend fun submitCheckIn(
         id: String,
         token: String,
-    ): Result<CheckInOut> = runApiCall { apiService.submitCheckIn(id = id, body = CheckInRequest(token = token)) }
+    ): Result<CheckInOut> =
+        runApiCall(mapHttpException = ::mapAuthAwareException) {
+            apiService.submitCheckIn(id = id, body = CheckInRequest(token = token))
+        }
 
     override suspend fun searchRegistrations(
         query: String,
         eventId: String?,
         authToken: String,
     ): Result<List<CheckInGuestOut>> =
-        runApiCall {
+        runApiCall(mapHttpException = ::mapAuthAwareException) {
             apiService.searchRegistrations(
                 auth = bearer(authToken),
                 query = query.takeIf { it.isNotBlank() },
@@ -53,31 +57,8 @@ class DefaultCheckInRepository(
             )
         }
 
-    private suspend fun <T> runApiCall(block: suspend () -> T): Result<T> =
-        try {
-            Result.success(block())
-        } catch (exception: HttpException) {
-            if (exception.code() == 401) {
-                Result.failure(UnauthorizedException())
-            } else {
-                Result.failure(ApiException(exception.message(), exception))
-            }
-        } catch (exception: SSLPeerUnverifiedException) {
-            Result.failure(
-                ApiException("Secure connection to the server could not be verified. Please update the app.", exception),
-            )
-        } catch (exception: IOException) {
-            Result.failure(ApiException("Unable to reach the server. Check your connection and try again.", exception))
-        } catch (exception: Exception) {
-            Result.failure(ApiException(exception.message ?: "Something went wrong.", exception))
-        }
+    private fun mapAuthAwareException(exception: HttpException): Exception =
+        if (exception.code() == 401) UnauthorizedException() else ApiException(exception.message(), exception)
 }
-
-class ApiException(
-    message: String,
-    cause: Throwable? = null,
-) : Exception(message, cause)
-
-class UnauthorizedException : Exception("You are not authorized to perform this action.")
 
 private fun bearer(token: String): String = "Bearer $token"
