+++ android/app/src/main/kotlin/be/champagnefestival/android/data/repository/EditionRepository.kt
@@ -3,9 +3,6 @@ package be.champagnefestival.android.data.repository
 import be.champagnefestival.android.data.api.ChampagneApiService
 import be.champagnefestival.android.data.model.EditionOut
 import be.champagnefestival.android.data.model.EventSummary
-import retrofit2.HttpException
-import java.io.IOException
-import javax.net.ssl.SSLPeerUnverifiedException
 
 interface EditionRepository {
     suspend fun getActiveEdition(): Result<EditionOut>
@@ -27,20 +24,5 @@ class DefaultEditionRepository(
     ): Result<List<EventSummary>> =
         runApiCall {
             apiService.getEvents(auth = "Bearer $authToken", editionId = editionId)
-        }
-
-    private suspend fun <T> runApiCall(block: suspend () -> T): Result<T> =
-        try {
-            Result.success(block())
-        } catch (exception: HttpException) {
-            Result.failure(ApiException(exception.message(), exception))
-        } catch (exception: SSLPeerUnverifiedException) {
-            Result.failure(
-                ApiException("Secure connection to the server could not be verified. Please update the app.", exception),
-            )
-        } catch (exception: IOException) {
-            Result.failure(ApiException("Unable to reach the server. Check your connection and try again.", exception))
-        } catch (exception: Exception) {
-            Result.failure(ApiException(exception.message ?: "Something went wrong.", exception))
         }
 }
